refactor: redefine how filler and listeners are organized
