--- typescript/solver/listeners/index.ts
@@ -1,1 +1,0 @@
-export { onChain } from "./onChain";
