+++ typescript/solver/solvers/onChain/listener.ts
@@ -1,10 +1,10 @@
 import { chainMetadata } from "@hyperlane-xyz/registry";
 import { MultiProvider } from "@hyperlane-xyz/sdk";
 
-import { OriginSettler__factory } from "../contracts/typechain/factories/OriginSettler__factory";
-import type { OpenEventArgs } from "../types";
+import { OriginSettler__factory } from "../../contracts/typechain/factories/OriginSettler__factory";
+import type { OpenEventArgs } from "../../types";
 
-const create = () => {
+export const create = () => {
   const { settlerContract } = setup();
 
   return function onChain(handler: (openEventArgs: OpenEventArgs) => void) {
@@ -32,5 +32,3 @@ function setup() {
 
   return { settlerContract };
 }
-
-export const onChain = { create };
